backend/requirements.txt
Flask
flask-cors
gspread
oauth2client
Flask-Mail
bcrypt

Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
gspread==5.10.0
oauth2client==4.1.3
bcrypt==4.0.1
python-dotenv==1.0.0
re==2.2.1